renames config.echo package to config.service
