+++ envoy-control-tests/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/config/service/EchoServiceExtension.kt
@@ -1,3 +1,3 @@
-package pl.allegro.tech.servicemesh.envoycontrol.config.echo
+package pl.allegro.tech.servicemesh.envoycontrol.config.service
 
 class EchoServiceExtension : ServiceExtension<EchoContainer>(EchoContainer())
